Fix Network Work: Change type in response model

+++ data/src/main/java/com/sdimosikvip/data/network/models/StockCompanyResponse.kt
@@ -17,7 +17,7 @@ data class StockCompanyResponse(
     @SerializedName("logo")
     val logo: String,
     @SerializedName("marketCapitalization")
-    val marketCapitalization: Int,
+    val marketCapitalization: Double,
     @SerializedName("name")
     val name: String,
     @SerializedName("phone")
